Coley: Added 2D Heat Conduction code (IGA2D)
An ugly, ugly port of the 2D heat conduction code.  Currently all inputs
for the problem are hard-coded into the code and the code assembles the
system matrix and forcing vector but does not actually solve the system.
It's ugly, but it works.  Posted here as a reference for others.